requirements.txt
python-telegram-bot>=20.7
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
tortoise-orm>=0.20.0
alembic>=1.12.1
aerich>=0.7.2
python-dotenv>=1.0.0
redis>=5.0.1
requests>=2.31.0
aiohttp>=3.9.1
aiosqlite>=0.19.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
loguru>=0.7.2
chatgpt-md-converter
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
jinja2>=3.1.3
